Refactor useMetadataFieldConfigs for reuse
