+++ context/app/static/js/components/entity-search/sidebar/ConfigureSearchTable/utils.js
