+++ context/app/static/js/components/entity-search/sidebar/ConfigureSearchTable/utils.spec.js
