+++ context/app/static/js/components/entity-search/sidebar/DonorSampleConfigureSearchTable/hooks.js
@@ -1,36 +1,9 @@
-import produce from 'immer';
-
-import metadataFieldtoEntityMap from 'metadata-field-entities';
+import { useMetadataFieldConfigs } from 'js/components/entity-search/sidebar/ConfigureSearchTable/hooks';
 import { useStore } from 'js/components/entity-search/SearchWrapper/store';
-import { createFacet } from 'js/components/entity-search/SearchWrapper/utils';
-import { getMetadataFieldsSortedByEntityTypeThenFieldName } from './utils';
 
-const relatedEntityTypesMap = {
-  donor: ['donor'],
-  sample: ['donor', 'sample'],
-  dataset: ['donor', 'sample'],
-};
-
-const sortedMetadataFields = getMetadataFieldsSortedByEntityTypeThenFieldName(metadataFieldtoEntityMap);
-
-function useMetadataFieldConfigs() {
-  const { entityType, initialFields, initialFacets } = useStore();
-  return sortedMetadataFields.reduce(
-    (acc, { fieldName, fieldEntityType }) => {
-      if (relatedEntityTypesMap[entityType].includes(fieldEntityType)) {
-        return produce(acc, (draft) => {
-          const group = `${fieldEntityType} Metadata`;
-          // eslint-disable-next-line no-param-reassign
-          return {
-            ...draft,
-            ...createFacet({ fieldName, entityType, facetGroup: group }),
-          };
-        });
-      }
-      return acc;
-    },
-    { ...initialFields, ...initialFacets },
-  );
+function useDonorSampleMetadatFieldConfigs() {
+  const { initialFields, initialFacets } = useStore();
+  return useMetadataFieldConfigs({ ...initialFields, ...initialFacets });
 }
 
-export { useMetadataFieldConfigs };
+export { useDonorSampleMetadatFieldConfigs };
